requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
dataclasses-json==0.5.7
Flask==2.3.2
frozenlist==1.3.3
google-search-results==2.4.2
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
langchain==0.0.180
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.24.3
openai==0.27.7
openapi-schema-pydantic==1.2.4
packaging==23.1
pydantic==1.10.8
python-dotenv==1.0.0
PyYAML==6.0
requests==2.31.0
SQLAlchemy==2.0.15
tenacity==8.2.2
tqdm==4.65.0
typing-extensions==4.6.2
typing-inspect==0.9.0
urllib3==1.26.6
Werkzeug==2.3.4
yarl==1.9.2
zipp==3.15.0
